feat: add query.control.ts to manage auth query keys and functions with unified QUERY_KEYS enum
